feat(airc-core): Identity struct + modular split (one concern per file) (#654)
* feat(airc-core): add Identity struct mirroring Python airc identity fields

PR-1 of the airc-core refactor stack per AIRC-RUST-SUBSTRATE.md +
the audit broadcast on #cambriantech.

The existing airc-core had bare newtype IDs (PeerId, ClientId, etc.)
but no Identity struct carrying the rich display fields the Python+bash
airc has been using: name, pronouns, role, bio, status, fingerprint,
integrations.

This commit ports the existing shape one-to-one — same six fields,
same defaults, same "/away" semantics, same UX-prompt heuristic
(is_complete = pronouns + role + bio all set). Per Joel's direction:
don't redesign, just port; we figured most of this out in the current
version.

API:
  - Identity::new(name) — fresh identity with all other fields default
  - Identity::is_complete() — UX heuristic for "should we nudge user
    to fill in pronouns/role/bio"
  - Identity::set_status(s) / .is_away() — IRC /away-style lifecycle
  - Serde derives — forward-compat (missing fields default to empty
    rather than fail deserialization)
  - integrations: BTreeMap<String, String> — consumer-registered
    cross-system identity binding (continuum persona ref, OpenClaw
    user id, GitHub login, etc.). airc never interprets values; just
    persists + transports. Deterministic key order via BTreeMap so
    serde encoding is stable across consumers.

Tests: 5 new cases:
  - new() sets name + defaults other fields
  - is_complete() requires pronouns + role + bio (mirrors Python heuristic)
  - away-status lifecycle including empty-string clears
  - serde round-trip with backward-compat (only-nick input deserializes)
  - integrations map handles multi-consumer namespacing without collision

9/9 airc-core tests pass total (4 existing + 5 new).

This is the first of ~7-8 small atomic PRs landing the audit gaps
called out on #cambriantech. Next up: Body enum (Json|Binary), then
EventFrame distinction, MediaRef split, Subscription, Control variants,
module split.

Targets `airc-rust` integration branch, not canary directly, per
Joel's branch-shape direction.


* refactor(airc-core): split lib.rs into per-concern modules

Per Joel's direction: "good modules, broken down into concerns / as
small a file as possible from the start / many files per module".

The previous lib.rs had grown to 490 LOC after the Identity addition.
Splitting it now (before more types land) so the small-file discipline
is baked in from the start of the airc-rust phase, not retrofitted
after the file balloons. CBAR substrate doctrine applied to airc-core
itself.

Module layout — one concern per file:

  - src/lib.rs        —  42 LOC, just module declarations + re-exports
  - src/ids.rs        —  50 LOC, newtype identifier wrappers
  - src/identity.rs   — 186 LOC, Identity struct + tests
  - src/transcript.rs —  92 LOC, TranscriptEvent + TranscriptKind + MentionTarget
  - src/cursor.rs     — 175 LOC, paging cursors + page_recent / page_before
  - src/receipt.rs    —  40 LOC, ReceiptKind + Receipt
  - src/attachment.rs —  72 LOC, AttachmentManifest
  - src/filter.rs     — 105 LOC, SelfFilter + filter_self_echoes

Largest module is identity.rs at 186 LOC (the most semantic surface);
median is ~92 LOC. All under the "thousands of lines" anti-pattern the
CBAR-substrate doctrine warns against.

Public API surface unchanged — every type that was previously at the
crate root is re-exported from lib.rs so `use airc_core::Identity;`
keeps working. Internal cross-module references use explicit module
paths (e.g. `crate::ids::PeerId`) so refactor visibility is preserved.

9/9 existing tests pass after the split. Tests move with their types
(Identity tests in identity.rs, paging tests in cursor.rs, filter tests
in filter.rs, attachment tests in attachment.rs).


---------
